cedict.el: Try to ignore "low-quality" definitions
Although in most cases multiple definitions for a term are grouped
into a single entry, CEDICT occasionally has multiple entries for the
same term.

By default our search process will return the first entry in such
cases, so when we find we're looking at a low-quality definition, try
to continue searching further for something better.

Currently the only definitions classified as "low-quality" are
/surname .../ entries.