cedict.el: Rewrite using a secondary data structure
The old lookup mechanism was purely based on buffer searches, and worked
reasonably well in many cases.  However there are significant number of
entries in CEDICT for which that mechanism doesn't work correctly, and
it's hard to really do things correctly while still being efficient
enough for interactive use.

This rewrite changes the lookup mechanism completely to incrementally
build a secondary lookup structure, which is used for the actual
lookups.  It gives more accurate lookup results, especially in some edge
cases, and should overall be faster than the old search-based
mechanism (it can be slightly slower the first time a word beginning
with a given character is looked up).

See the header comment in cedict.el for details.